successfully made the build infrastructure {build / be capable of building} separate "badblocks_with_crypto" and "badblocks" executables...
... so that OS package maintainers will still have the same old "badblocks" executable that they are used to,
with the same run-time dependencies (i.e. shared library requirements) as before.  In other words,
the "badblocks" program should _always_ be usable, even in situations in which "badblocks_with_crypto"
is unusable for some reason that is not controllable at the source-code level (e.g. lack of shared libraries
on an emergency boot medium such as a "rescue disk").