a slight functional change -- and _massive_ re-indentation -- of most of the code I have added to "/configure.ac"...
... since I started working on this codebase; the objective of this change: to not even try to _configure_ the crypto build
when it is not the case that _both_ of the currently-required inputs to "<...>/configure" have been provided.  As of now,
on my dev. system, the crypto build files [i.e. "badblocks_with_crypto.o" & "badblocks_with_crypto" [the executable] are
built _anyway_ -- but without the actual crypto code in them!  [this bug probably existed prior to this commit, but I don`t
really know for sure right now]