_lots_ of work in/on "/configure.ac" [regenerated "/configure"], plus a little bit of work in/on "/misc/Makefile.in"...
... with the end result[s] that at least now in a non-{crypto-enabled} build the relevant build is neither compiled nor linked
_by_ _default_, although -- since this was accomplished via nasty hacks and dirty kludges -- you _can_ force building to fail
unexpectedly by doing a non-crypto "configure" followed by something like "make badblocks_with_crypto.o" or
"badblocks_with_crypto" if you really _want_ to do so.

Also made lots of progress in re adding code to "/configure.ac" for adopting the *BSD compatibility library [for the purpose
of "stealing" their seemingly-superbly-well-suited-to-the-current-need "badblocks_with_crypto" function]: AFAIK the only part
of this part [the "/configure.ac" part] that is not yet done is the "can a test program be compiled and linked" part.